requirements.txt
streamlit
pandas
numpy
python-dateutil
sentence-transformers
faiss-cpu
torch
pyarrow
transformers==4.41.2
accelerate